Move non-controller code in 'module_utils/cm_utils' into new 'cm_controller_utils.py' file
